# tickets-module-api/resources/api.yaml
openapi: 3.0.1
info:
  title: Сервис продажи билетов
  version: 1.0.0
servers: [ ]
tags:
  - name: TicketsResource
    description: "API для работы с билетами"
paths:
  /login:
    post:
      description: Аутентификация пользователя
      operationId: login
      tags:
        - OauthController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/LoginRequest"
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/LoginResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        404:
          description: Объект не найден
        500:
          description: Ошибка сервера/сервиса/БД
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"

  /token:
    post:
      description: Обновление access токена по refresh токену
      operationId: refreshToken
      tags:
        - OauthController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/RefreshTokenRequest"
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/LoginResponse"
        401:
          description: Невалидный refresh token

  /register:
    post:
      description: Регистрация пользователя
      operationId: register
      tags:
        - OauthController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/RegisterRequest"
      responses:
        201:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/LoginResponse"

  /carriers:
    post:
      description: Метод создания компании перевозчика
      operationId: createCarrier
      tags:
        - CarriersController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/CreateCarrierRequest"
      responses:
        201:
          description: Компания перевозчик успешно сохранена
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateCarrierResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        500:
          description: Внутренняя ошибка сервера
    get:
      description: Метод предоставления списка компаний перевозчиков
      operationId: getAllCarriers
      tags:
        - CarriersController
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: "types.yaml#/components/schemas/Pageable"
        - in: query
          name: searchParam
          required: false
          schema:
            $ref: "types.yaml#/components/schemas/CarriersSearchRequest"
      responses:
        200:
          description: Поиск выполнен успешно
          headers:
            X-Total-Count:
              description: Общее количество элементов при постраничном отображении
              schema:
                type: integer
                format: int32
                example: 100
            Link:
              description: Ссылки на страницы
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/Page"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        404:
          description: Объекты не найдены
        500:
          description: Внутренняя ошибка сервера

  /carriers/{id}:
    get:
      description: Метод получения компании перевозчика по id
      operationId: getCarrier
      tags:
        - CarriersController
      parameters:
        - name: id
          in: path
          description: Идентификатор компании перевозчика
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Операция обработана успешно, результат предоставлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateCarrierResponse"
        400:
          description: Переданы не корректные параметры запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        404:
          description: Объект не найден
        500:
          description: Ошибка сервера/сервиса/БД
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"

    patch:
      description: Частичное обновление данных компании перевозчика по id
      operationId: patchCarrier
      tags:
        - CarriersController
      parameters:
        - name: id
          in: path
          description: Идентификатор компании перевозчика
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "patch-types.yaml#/components/schemas/CarrierPatchRequest"
      responses:
        201:
          description: Компания перевозчик успешно обновлена
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateCarrierResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей

  /routes:
    post:
      description: Метод создания маршрута
      operationId: createRoute
      tags:
        - RoutesController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/CreateRouteRequest"
      responses:
        201:
          description: Маршрут успешно сохранен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateRouteResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        500:
          description: Внутренняя ошибка сервера
    get:
      description: Метод предоставления списка маршрутов
      operationId: getAllRoutes
      tags:
        - RoutesController
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: "types.yaml#/components/schemas/Pageable"
        - in: query
          name: searchParam
          required: false
          schema:
            $ref: "types.yaml#/components/schemas/RoutesSearchRequest"
      responses:
        200:
          description: Поиск выполнен успешно
          headers:
            X-Total-Count:
              description: Общее количество элементов при постраничном отображении
              schema:
                type: integer
                format: int32
                example: 100
            Link:
              description: Ссылки на страницы
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/Page"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        404:
          description: Объекты не найдены
        500:
          description: Внутренняя ошибка сервера

  /routes/{id}:
    get:
      description: Метод получения маршрута по id
      operationId: getRoute
      tags:
        - RoutesController
      parameters:
        - name: id
          in: path
          description: Идентификатор маршрута
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Операция обработана успешно, результат предоставлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateRouteResponse"
        400:
          description: Переданы не корректные параметры запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        404:
          description: Объект не найден
        500:
          description: Ошибка сервера/сервиса/БД
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"

    patch:
      description: Частичное обновление данных маршрута по id
      operationId: patchRoute
      tags:
        - RoutesController
      parameters:
        - name: id
          in: path
          description: Идентификатор маршрута
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "patch-types.yaml#/components/schemas/RoutePatchRequest"
      responses:
        201:
          description: Маршрут успешно обновлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateRouteResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей

  /tickets:
    post:
      description: Метод создания билета
      operationId: createTicket
      tags:
        - TicketsController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/CreateTicketRequest"
      responses:
        201:
          description: Билет успешно сохранен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateTicketResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        500:
          description: Внутренняя ошибка сервера
    get:
      description: Метод предоставления списка билетов
      operationId: getAllTickets
      tags:
        - TicketsController
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: "types.yaml#/components/schemas/Pageable"
        - in: query
          name: searchParam
          required: false
          schema:
            $ref: "types.yaml#/components/schemas/TicketsSearchRequest"
      responses:
        200:
          description: Поиск выполнен успешно
          headers:
            X-Total-Count:
              description: Общее количество элементов при постраничном отображении
              schema:
                type: integer
                format: int32
                example: 100
            Link:
              description: Ссылки на страницы
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/Page"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        404:
          description: Объекты не найдены
        500:
          description: Внутренняя ошибка сервера

  /tickets/{id}:
    get:
      description: Метод получения билета по id
      operationId: getTicket
      tags:
        - TicketsController
      parameters:
        - name: id
          in: path
          description: Идентификатор билета
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Операция обработана успешно, результат предоставлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateTicketResponse"
        400:
          description: Переданы не корректные параметры запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        404:
          description: Объект не найден
        500:
          description: Ошибка сервера/сервиса/БД
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"

    patch:
      description: Частичное обновление данных билета по id
      operationId: patchTicket
      tags:
        - TicketsController
      parameters:
        - name: id
          in: path
          description: Идентификатор билета
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "patch-types.yaml#/components/schemas/TicketPatchRequest"
      responses:
        201:
          description: Маршрут успешно обновлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateTicketResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей

  /tickets/{id}/buy:
    post:
      description: Метод покупки билета по id
      operationId: buyTicket
      tags:
        - TicketsController
      parameters:
        - name: id
          in: path
          description: Идентификатор билета
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Операция обработана успешно, результат предоставлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateTicketResponse"
        400:
          description: Переданы не корректные параметры запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        404:
          description: Объект не найден
        500:
          description: Ошибка сервера/сервиса/БД
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"

  /tickets/{id}/return:
    post:
      description: Метод возврата билета по id
      operationId: returnTicket
      tags:
        - TicketsController
      parameters:
        - name: id
          in: path
          description: Идентификатор билета
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Операция обработана успешно, результат предоставлен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/CreateTicketResponse"
        400:
          description: Переданы не корректные параметры запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        404:
          description: Объект не найден
        500:
          description: Ошибка сервера/сервиса/БД
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"

  /tickets/release-batch:
    post:
      description: Метод создания билета
      operationId: releaseTicketsBatch
      tags:
        - TicketsController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "types.yaml#/components/schemas/ReleaseTicketsBatchRequest"
      responses:
        201:
          description: Билет успешно сохранен
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ReleaseTicketsBatchResponse"
        400:
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: "types.yaml#/components/schemas/ErrorResponse"
        401:
          description: Отсутствует или невалиден JWT-токен
        403:
          description: Пользователь не имеет ни одной из необходимых ролей
        500:
          description: Внутренняя ошибка сервера